Add missing cast for PSP compatibility

+++ src/st/get_normalized_angle.h
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 u16 GetNormalizedAngle(u16 arg0, u16 arg1, u16 arg2) {
-    u16 temp_a2 = arg2 - arg1;
+    u16 temp_a2 = (s16)(arg2 - arg1);
     u16 ret;
 
     if (temp_a2 & 0x800) {
